test: Ajusta comparação do AssertTrue da variável ufBrasil no teste unitário da consulta cadastro contribuinte da NFCe. ID #173318

+++ source/Unimake.DFe.Test/NFCe/ConsultaCadastroTest.cs
@@ -122,7 +122,7 @@ namespace Unimake.DFe.Test.NFCe
             var consultaCadastro = new ConsultaCadastro(ufBrasil, cnpj, configuracao);
             consultaCadastro.Executar();
 
-            Assert.True(configuracao.CodigoUF.Equals(ufBrasil), "UF definida nas configurações diferente de " + ufBrasil.ToString());
+            Assert.True(configuracao.CodigoUF.Equals((int)ufBrasil), "UF definida nas configurações diferente de " + ufBrasil.ToString());
             Assert.True(consultaCadastro.Result.InfCons.CUF.Equals(ufBrasil), "Webservice retornou uma UF e está diferente de " + ufBrasil.ToString());
             Assert.True(consultaCadastro.Result.InfCons.CStat != 259, "CNPJ consultado não é foi localizado no webservice da UF " + ufBrasil.ToString() + ".");
         }
